refactor: 이미지 경로 설정 변경
