+++ frontend/package.json
@@ -10,7 +10,6 @@
     "test": "jest"
   },
   "dependencies": {
-    "@gorhom/bottom-sheet": "^4.4.5",
     "@gorhom/bottom-sheet": "^5.1.4",
     "@react-native-async-storage/async-storage": "^2.1.2",
     "@react-native-community/checkbox": "^0.5.17",
@@ -22,8 +21,8 @@
     "react": "^19.0.0",
     "react-native": "^0.79.2",
     "react-native-calendars": "^1.1312.0",
-    "react-native-elements": "^3.4.3",
     "react-native-date-picker": "^5.0.12",
+    "react-native-elements": "^3.4.3",
     "react-native-fast-image": "^8.6.3",
     "react-native-gesture-handler": "^2.25.0",
     "react-native-image-picker": "^8.2.1",
@@ -33,10 +32,10 @@
     "react-native-responsive-screen": "^1.4.2",
     "react-native-safe-area-context": "^5.4.0",
     "react-native-screens": "^4.10.0",
-    "react-native-swipe-list-view": "^3.2.9",
     "react-native-shadow-2": "^7.1.1",
     "react-native-share": "^12.0.11",
     "react-native-svg": "^15.12.0",
+    "react-native-swipe-list-view": "^3.2.9",
     "react-native-vector-icons": "^10.2.0",
     "react-native-view-shot": "^4.0.3",
     "react-native-webview": "^13.13.5"
