+++ frontend/src/components/layouts/MainHeader.tsx
@@ -4,8 +4,8 @@ import { View, Image, StyleSheet } from 'react-native';
 export default function Header() {
   return (
     <View style={styles.header}>
-      <Image source={require('../assets/icons/appLogo.png')} style={styles.logo} />
-      <Image source={require('../assets/icons/appTitle.png')} style={styles.title} />
+      <Image source={require('../../assets/icons/appLogo.png')} style={styles.logo} />
+      <Image source={require('../../assets/icons/appTitle.png')} style={styles.title} />
     </View>
   );
 }
