+++ frontend/src/screens/sharing/SharingMain.tsx
@@ -14,8 +14,8 @@ import {
 import { useNavigation } from '@react-navigation/native';
 import { widthPercentageToDP as wp, heightPercentageToDP as hp } from 'react-native-responsive-screen';
 import type { NativeStackNavigationProp } from '@react-navigation/native-stack';
-import Footer from "../../components/Footer";
-import CustomSearchBar from "../../components/CustomSearchBar";
+import Footer from "../../components/layouts/Footer";
+import CustomSearchBar from "../../components/common/CustomSearchBar";
 import { SharingStackParamList } from "../../navigation/SharingNavigator";
 
 export default function ShairingMain() {
